% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareindex.R
\name{rareindex}
\alias{rareindex}
\title{calculates the rr index}
\usage{
rareindex(table)
}
\arguments{
\item{table}{containing the information necessary to calculate index of rarity}
}
\value{
returns a matrix containing the values of 4 indexes-gri,hsi,psi and rr
}
\description{
Function to calculate a index of rarity based on Rabinowitz scheme
sp Prepare a list of unique species
lat1  Calculates a latitudinal amplitude for each species
long1 Calculates a longitudinal amplitude for each species
latlong1 sum the lat1 and long1
range1 Adds constant 1 to the latlong1
max_ind calculate the maximum population size for each species
Habitat calculate the maximum number of habitat where species are for each species
rareindex Returns the rarity index
}
\references{
Rabinowitz 1981 - Seven forms of rarity. In 'The biological aspects of rare plant conservation'.Ed. H Synge pp. 205-217
}
\author{
Everton Alves Maciel
}
