% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareplot.R
\name{rareplot}
\alias{rareplot}
\title{Creates 4 plots with the rare indexes}
\usage{
rareplot(table)
}
\arguments{
\item{table}{containing the information necessary to plot}
}
\description{
function to plot index values on a 4 charts
}
\details{
The rrplot uses the table function to read and manipulate the data

use the par to create a 2 x 2 layout,

then the graph is used to create four separate graphs
}
\author{
Everton Alves Maciel
}
