% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrboxplot.R
\name{rrboxplot}
\alias{rrboxplot}
\title{Creates a boxplot containing the averages of four indexes}
\usage{
rrboxplot(table)
}
\arguments{
\item{table}{containing the information necessary to plot}
}
\description{
function to plot index values on a single charts
}
\details{
The rrboxplot function uses the function table to create 4 box in a single layout;

The rows in each box represent the average of each index
}
\author{
Everton Alves Maciel
}
